# public/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Quran Api service for mobile App",
    "version": "0.0.1",
    "title": "Alquran  API",
    "termsOfService": "",
    "contact": {
      "email": "abdo.gamy2010@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "quran.uniative.com",
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "securityDefinitions": {
    "api_key": {
      "type": "apiKey",
      "name": "api_token",
      "in": "header"
    }
  },
  "paths": {
...
    "/contests/details": {
      "get": {
        "tags": [
          "Contests"
        ],
        "summary": "Contest details",
        "description": "Contest details",
        "operationId": "contests.details",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "query",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "contest_id": {
                  "type": "integer"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/ContestWithMembers"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "first_name": {
          "type": "string"
        },
        "last_name": {
          "type": "string"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        },
        "photo_id": {
          "type": "integer"
        },
        "last_login": {
          "type": "integer"
        },
        "role_id": {
          "type": "integer"
        },
        "status": {
          "type": "integer"
        },
        "lang": {
          "type": "string"
        },
        "color": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "about": {
          "type": "string"
        },
        "google_plus": {
          "type": "string"
        },
        "twitter": {
          "type": "string"
        },
        "facebook": {
          "type": "string"
        },
        "photo": {
          "$ref": "#/definitions/Media"
        }
      }
    },
    "Media": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "type": {
          "type": "string"
        },
        "path": {
          "type": "string"
        },
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "provider": {
          "type": "string"
        },
        "provider_id": {
          "type": "string"
        },
        "provider_image": {
          "type": "string"
        },
        "length": {
          "type": "integer"
        },
        "created_at": {
          "type": "string"
        },
        "updated_at": {
          "type": "string"
        }
      }
    },
...
    "ContestWithMembers": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer"
        },
        "name": {
          "type": "string"
        },
        "goal": {
          "type": "string"
        },
        "start_at": {
          "type": "string"
        },
        "expired_at": {
          "type": "string"
        },
        "user_id": {
          "type": "integer"
        },
        "is_expired": {
          "type": "boolean"
        },
        "is_joined": {
          "type": "boolean"
        },
        "is_opened": {
          "type": "boolean"
        },
        "member_counter": {
          "type": "integer"
        },
        "winner_id": {
          "type": "integer"
        },
        "members": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        },
        "creator": {
          "$ref": "#/definitions/User"
        },
        "winner": {
          "$ref": "#/definitions/User"
        }
      }
    },
...
  }
}